feat(GBGB-268-fe-refec-all) - 전체 코드 리펙토링 완료, 코드 정리, 모달 분리 등 (마이페이지 제외) - 온보딩, 경기장 정보 선택 수정 (익사이팅존, 프리미엄 추가)
