+++ goorm-gongbang/components/my/StadiumMapData.ts
@@ -297,11 +297,13 @@ export const STADIUM_PATHS = [
     },
     {
         "d": "M173.942 239.911C181.408 257.829 213.264 301.133 213.264 301.133L207.809 326.118C191.309 304 173.345 273 151.544 236.925L149.877 235.535L173.942 239.911Z",
-        "fill": "#6D6D6D"
+        "fill": "#6D6D6D",
+        "num": 3
     },
     {
         "d": "M366.309 435C366.309 435 348.309 440.5 319.309 440.5C290.309 440.5 272.309 435 272.309 435L283.809 388.5C289.975 390.833 305.709 394.2 319.309 394.2C332.909 394.2 348.642 390.833 354.809 388.5L366.309 435Z",
-        "fill": "#0460DA"
+        "fill": "#0460DA",
+        "num": 1
     },
     {
         "d": "M265.963 381.604C270.859 383.785 276.053 385.751 281.604 387.506L277.333 401.896C271.755 399.744 266 397.189 260.132 394.2L265.963 381.604Z",
@@ -545,7 +547,8 @@ export const STADIUM_PATHS = [
     },
     {
         "d": "M465.63 239.911C458.164 257.829 426.309 301.133 426.309 301.133L431.764 326.118C448.264 304 466.227 273 488.028 236.925L489.695 235.535L465.63 239.911Z",
-        "fill": "#6D6D6D"
+        "fill": "#6D6D6D",
+        "num": 2
     },
     {
         "d": "M466 39L476.5 25C489.7 33 508.667 50.6667 516.5 58.5L504.5 72C494.1 60 474.5 45 466 39Z",
